Merged PR 16: change logger to logging to avoid logger not defined error
We find that in job-exporter, the error "logger is not defined" in container_ib_stat.py will show up and cause some of the IB related metrics missing. We haven't found the root cause of this problem yet but have a workaround to remove the errors with using logging.* derectly instead of logger. This PR is a hotfix to make the job exporter working properly with this change.