Merged PR 23: update the local storage mount to multiple places
After discussed with Ziyue, we create this PR to update the local storage mount to multiple places. For backward compatibility, we follow the original design but extend it. The multiple paths are separated by the character ":". For instance:
mntpath: "/path1:/path2:/path3"
hostpath: "/hostpath1:/hostpath2:/hostpath3"

and the system will mount them like:
{hostpath: /hostpath1, mntpath: /path1}
{hostpath: /hostpath2, mntpath: /path2}
{hostpath: /hostpath3, mntpath: /path3}

If the path number of mntpath and hostpath are not equal, or some paths are not Linux path style, the job won't be created.

The code has been tested on PAI-TEST cluster.