Merged PR 69: Fix the user's group list when changing group list
##[bugifx] Fix the user's group list when changing group list

### Description
When the group information is updated in rest-server, the user's related information cannot be updated unless the administrator updated it user by user manually. This PR is going to fix this problem. After testing, we find that the existing delete group REST API can remove the group information for each user, but the newly added group information cannot be updated for each user. To update the new-added group information, we add a REST API to clean all the tokens in the AKS system and add token reading from cache and token verifying in the code. After we removed all the tokens, the user needs to re-login to the service and at that time, the new group information will be updated into the user's account.

### Test Actions
1. List all users to check their group list
2. Call REST API remove a group from the system which is in the tester's group list
3. List all users to check if the group has been removed from the user information as well as the VCs associated with this group
4. Add new group into the system
5. List all users to check their group list again
6. Call REST API to remove all the tokens in the system
7. Login to the system again
8. List all users to check if the group list has been updated

### Expected Results
1. All the user's information including the group list for each user
2. The API returned status code 200
3. All the user information has been updated. The removed group and its associated VCs won't been seen in the result.
4. The service restarts successfully
5. All the user information is listed, as same as step 3.
6. The API returned status code 200
7. The message box "token is invalid" is showed, and after login, the main page can be seen
8. All the user information is listed, and the new group and VCs have been added in the tester's user information.